todo-app/pyproject.toml
[project]
name = "todo-app"
version = "0.1.0"
description = "FastAPI Todo application with SQLModel and PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlmodel",
    "psycopg2-binary",
    "python-dotenv",
    "pytest",
    "httpx",
    "pwdlib[argon2]>=0.3.0",
    "python-jose[cryptography]>=3.5.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
]
